webhook/requirements.txt
fastapi
python-dotenv
asyncpg
uvicorn